requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiohttp-proxy==0.1.2
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
lxml==4.9.3
magic-filter==1.0.12
multidict==6.0.4
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.2
